% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_soothsayer.R
\name{add_soothsayer}
\alias{add_soothsayer}
\title{Create soothsayer from existing models}
\usage{
add_soothsayer(mable, ...)
}
\arguments{
\item{mable}{The mable to add soothsayer to.}

\item{...}{The soothsayer to add - supports multiple soothsayers.}
}
\value{
A mable.
}
\description{
Add a soothsayer to an existing **mable** wthout reestimating models which are available.
}
\details{
This is intended to steamline adding soothsayer to existing mables - either as a model selector,
as ensembler, or anything else you wish. The models are not reestimated - but the missing models should be (although they
are only added to soothsayer currently, not the baseline mable).
}
\note{
Might be broken for multiple soothsayer/time series combinations.
}
