% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combiners.R
\name{combiner_greedy_stacking}
\alias{combiner_greedy_stacking}
\alias{combiner_equal}
\alias{combiner_custom}
\alias{combiner_oracle}
\alias{combiner_lm}
\alias{combiner_after}
\title{Combination functions which return combination weights}
\usage{
combiner_greedy_stacking(
  .models,
  prior_weights = NULL,
  metric = rmse_stack,
  oracle_weights = NULL,
  ...
)

combiner_equal(.models, prior_weights = NULL, oracle_weights = NULL, ...)

combiner_custom(.models, prior_weights = NULL, ...)

combiner_oracle(.models, prior_weights = NULL, oracle_weights = NULL, ...)

combiner_lm(.models, prior_weights = NULL, oracle_weights = NULL, ...)

combiner_after(.models, prior_weights = NULL, ...)
}
\arguments{
\item{.models}{The fitted models - this is a named or unnamed list of fitted fable compatible models.}

\item{prior_weights}{Prior weights to use for combinations - see details.}

\item{metric}{The metric to use when using greedy stacking (defaults to rmse).}

\item{oracle_weights}{Weights returned by an oracle (if available).}

\item{...}{Additional parameters (currently not implemented).}
}
\value{
A vector of length .models containing the individual weights
}
\description{
Functions which calculate ensemble model combination weights
}
\details{
Prior weights can be passed to combiner_custom and they will be the
only weights computed by that combiner. Other combiners take the average of prior_weights
and whatever weights they output.
}
