% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_features.R
\name{compute_features}
\alias{compute_features}
\title{Compute features for a tsibble}
\usage{
compute_features(x, ...)
}
\arguments{
\item{x}{A tsibble.}

\item{...}{Additional arguments.}
}
\value{
A data.frame wth computed time series features, with ncol features and
nrow individual time series.
}
\description{
Compute features for a tsibble
}
