% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_accuracy.R
\name{rmsse}
\alias{rmsse}
\alias{mase}
\alias{mpe}
\alias{mape}
\alias{maape}
\alias{me}
\alias{rmse}
\alias{mae}
\alias{acf1}
\alias{fast_measure}
\alias{fast_accuracy}
\title{Fast accuracy measure computation}
\usage{
rmsse(resid_sqr, scale, na.rm = TRUE)

mase(resid, scale, na.rm = TRUE)

mpe(scl_prc_resid, na.rm = TRUE)

mape(scl_prc_resid, na.rm = TRUE)

maape(scl_prc_resid, na.rm = TRUE)

me(resid, na.rm = TRUE)

rmse(resid_sqr, na.rm = TRUE)

mae(resid, na.rm = TRUE)

acf1(resid, na.rm = TRUE, demean = TRUE)

fast_measure(resid, actual, scale = 1, na.rm = TRUE)

fast_accuracy(
  fable,
  test,
  train = NULL,
  measures = fast_measure,
  across = c(".model", "h"),
  ...
)
}
\arguments{
\item{resid_sqr}{Squared residuals}

\item{scale}{Scaling factor - typically mean of training data}

\item{na.rm}{Whether to remove NA values - defaults to **TRUE**}

\item{resid}{Residuals}

\item{scl_prc_resid}{Scaled percentage residuals}

\item{demean}{Whether to demean before calculating autocorrelations - defaults to **TRUE**}

\item{actual}{Actuals}

\item{fable}{The model fable - we do not work with mables}

\item{test}{The test data corresponding to forecasts in the fable}

\item{train}{The original data used in fitting the models - if available,
scaled accuracy measures will be computed by default.}

\item{measures}{Currently only for future extensions, though **fast_measure**
does everything that was required so far}

\item{across}{The grouping variables across which to compute measures - by default
just the ".model" column, but you can easily pass something like "h" and it will work.}

\item{...}{For future extensions :)}
}
\value{
A data.frame with computed accuracy measures, grouped by whatever was supplied in
**across** and the keys present in **test**.
}
\description{
A pretty fast way to compute commong accuracy measures
}
