% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_model_parameters.R
\name{fix_model_parameters}
\alias{fix_model_parameters}
\title{Create a fable model with fixed parameters}
\usage{
fix_model_parameters(model, ...)
}
\arguments{
\item{model}{The model to use - typically a **fable** model such as \link[fable]{AR}}

\item{...}{The arguments to pass, the RHS of the formula to be passed to the underlying model.}
}
\value{
A New, fable compatible model, with the parameters fixed.
}
\description{
Fix some parameters of a fable model
}
