% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{random_forecast_h}
\alias{random_forecast_h}
\alias{soothsayer_forecaster}
\title{Soothsayer oracle training utilities}
\usage{
random_forecast_h(x, min_h = 4, max_h = 12, ...)

soothsayer_forecaster(
  series,
  models = list(arima = fable::ARIMA, fable::SNAIVE, fable::RW, theta = fable::THETA,
    ets = fable::ETS, nnetar = fable::NNETAR, croston = fable::CROSTON, ar = fable::AR),
  forecast_h = random_forecast_h,
  save_forecast_experiment = TRUE,
  save_outfile = "1",
  save_folder = paste0("experiment_", Sys.Date()),
  save_n_threads = 12
)
}
\arguments{
\item{x}{The time series (a vector).}

\item{min_h}{The minimal length of the forecast}

\item{max_h}{The maximal length of the forecast}

\item{...}{Not implemented.}

\item{series}{The series to use for modelling, a \link[tsibble]{tsibble}. Note that this must only have a
single measure variable.}

\item{models}{A named list of fable compatible model functions, such as \link[fable]{ARIMA}.
This must be a function which returns a valid fable model definition, preferably generated by
\link[fabletools]{new_model_definition}}

\item{forecast_h}{The forecast length h, either an integer, or a function which can return one,
possibly based on the underlying time series. For example, see \link[soothsayer]{random_forecast_h}.}

\item{save_forecast_experiment}{Whether to save the forecasting experiment, by default **TRUE**, and uses
quicksave (see \link[qs]{qsave})}

\item{save_outfile}{An optional argument to add an identifier to the saved objects.}

\item{save_folder}{The folder to save the objects to.}

\item{save_n_threads}{The number of threads to use when saving objects, see documentation for \link[qs]{qsave}.}
}
\value{
A random forecast length, sampled uniformly between **min_h** and **max_h**.

A list with fitted models, generated forecasts and computed forecast accuracies.
}
\description{
Utilities for training oracle models

Utilities for training oracle models
}
