% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soothsayer.R
\name{soothsayer}
\alias{soothsayer}
\title{Soothsayer model}
\usage{
soothsayer(formula, ...)
}
\arguments{
\item{formula}{A soothsayer model formula (see details).}

\item{...}{Additional arguments (see details).}
}
\value{
A soothsayer model, analogous to other model objects within fable/fabletools.
}
\description{
The main function of soothsayer - creates soothsayer models trainable and usable within fable and fabletools.
}
\details{
Accepts and parses several model specials - notably rules, oracle, feature_set, model_aliases, combinator
}
\note{
Maybe some other day.
}
