% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soothsayer_booster.R
\name{soothsayer_booster}
\alias{soothsayer_booster}
\title{Soothsayer boosting model}
\usage{
soothsayer_booster(formula, ...)
}
\arguments{
\item{formula}{A soothsayer model formula (see details).}

\item{...}{Additional arguments (see details).}
}
\value{
A soothsayer model, analogous to other model objects within fable/fabletools.
}
\description{
Boosting - stacking multiple models by modelling the residuals of the first model by the second one.
}
\details{
Accepts and parses several model specials.
}
\note{
Maybe some other day.
}
