% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\docType{data}
\name{entropy}
\alias{entropy}
\alias{lumpiness}
\alias{stability}
\alias{nonlinearity}
\alias{hurst}
\alias{unitroot_pp}
\alias{box_pierce}
\alias{unitroot_kpss}
\alias{longest_flat_spot}
\alias{n_crossing_points}
\alias{ljung_box}
\alias{acf_features}
\alias{intermittent}
\alias{arch_stat}
\alias{spectral}
\alias{catch22_feat}
\alias{positive}
\alias{negative}
\alias{zeros}
\alias{continuous}
\alias{count}
\alias{period}
\alias{tslength}
\alias{boxcox_lambda}
\alias{unitroot_ndiffs}
\alias{feat_stl}
\alias{unitroot_nsdiffs}
\alias{shift_kl_max}
\alias{shift_level_max}
\alias{shift_var_max}
\alias{soothsayer_feature_set}
\title{Features used within soothsayer}
\format{
An object of class \code{list} of length 29.
}
\usage{
entropy(x)

lumpiness(x)

stability(x)

nonlinearity(x)

hurst(x)

unitroot_pp(x)

box_pierce(x)

unitroot_kpss(x)

longest_flat_spot(x)

n_crossing_points(x)

ljung_box(x)

acf_features(x)

intermittent(x)

arch_stat(x)

spectral(x)

catch22_feat(x)

positive(x)

negative(x)

zeros(x)

continuous(x)

count(x)

period(x)

tslength(x)

boxcox_lambda(x)

unitroot_ndiffs(x)

feat_stl(x)

unitroot_nsdiffs(x)

shift_kl_max(x)

shift_level_max(x)

shift_var_max(x)

soothsayer_feature_set
}
\arguments{
\item{x}{A vector to apply measures to.}
}
\description{
All of the features in soothsayer - most taken from other sources,
see **details**
}
\keyword{datasets}
