% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\docType{data}
\name{soothsayer_alias_set}
\alias{soothsayer_alias_set}
\title{Soothsayer model aliases}
\format{
An object of class \code{list} of length 36.
}
\usage{
soothsayer_alias_set
}
\description{
A set of soothsayer model aliases (for purposes of matching).
}
\keyword{datasets}
