% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{new_soothsayer_oracle}
\alias{new_soothsayer_oracle}
\alias{fit.soothsayer_oracle}
\alias{predict.soothsayer_oracle}
\title{Defne a soothsayer oracle}
\usage{
new_soothsayer_oracle(
  oracle_name = NULL,
  feature_data = NULL,
  series_data = NULL,
  feature_set = soothsayer_feature_set,
  forecast_accuracies = NULL,
  forecaster = soothsayer_forecaster,
  train = function(accuracies, features, ...) {    
    rlang::abort("This oracle must have a training method.") },
  predict = function(...) {    
    rlang::abort("This oracle must have a prediction method.") },
  emits = "models",
  ...
)

\method{fit}{soothsayer_oracle}(object, ...)

\method{predict}{soothsayer_oracle}(object, features, ...)
}
\arguments{
\item{oracle_name}{= NULL,}

\item{feature_data}{= NULL, # either the full feature data,}

\item{series_data}{= NULL, # or series data}

\item{feature_set}{= soothsayer_feature_set, # always a feature set - this is necessary as
this will be checked against}

\item{forecast_accuracies}{= NULL, # you can have forecast accuracies}

\item{forecaster}{= soothsayer_forecaster, # use our default forecastng engine}

\item{train}{The training function for your oracle.}

\item{predict}{The prediction for your oracle.}

\item{emits}{What the model emits (either **"models"** or **"weights"**) where **models**
is a variable length character vector, e.g. **c("ets","ar")**, or a named vector of weights,
e.g. **c("ets" = 0.3, "ar" = 0.7)**.}

\item{...}{Additional arguments (currently unimplemented).}

\item{object}{The soothsayer oracle to fit/use for prediction.}

\item{features}{The features to use for the predict function (analogous to newx/newdata in
other predict functions).}
}
\value{
A soothsayer oracle
}
\description{
Create a new soothsayer oracle
}
