% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{new_soothsayer_oracle}
\alias{new_soothsayer_oracle}
\alias{predict.soothsayer_oracle}
\title{Define a soothsayer oracle}
\usage{
new_soothsayer_oracle(
  oracle_name = NULL,
  trained_oracle = NULL,
  predict = function(...) {
    
    rlang::abort("This oracle must have a prediction method.")
 },
  emits = "models",
  ...
)

\method{predict}{soothsayer_oracle}(object, features, ...)
}
\arguments{
\item{oracle_name}{= NULL,}

\item{trained_oracle}{A trained oracle (if needed) for making predictions.}

\item{predict}{The prediction for your oracle.}

\item{emits}{What the model emits (either **"models"** or **"weights"**) where **models**
is a variable length character vector, e.g. **c("ets","ar")**, or a named vector of weights,
e.g. **c("ets" = 0.3, "ar" = 0.7)**.}

\item{...}{Additional arguments (currently unimplemented).}

\item{object}{The soothsayer oracle to fit/use for prediction.}

\item{features}{The features to use for the predict function (analogous to newx/newdata in
other predict functions).}
}
\value{
A soothsayer oracle
}
\description{
Create a new soothsayer oracle
}
